+++ test/utilities/error.ts
@@ -1,0 +1,3 @@
+export function customError(errorName: string): string {
+  return `VM Exception while processing transaction: reverted with custom error '${errorName}()'`;
+}
