+++ test/utilities/random.ts
@@ -1,5 +1,3 @@
-import { Variants } from "./interfaces";
-
 export function getRandom(rnd: () => number, min: number, max: number) {
   const minL = Math.log(min);
   const maxL = Math.log(max);
@@ -9,14 +7,14 @@ export function getRandom(rnd: () => number, min: number, max: number) {
   return res;
 }
 
-export function choice(rnd: () => number, obj: Variants) {
+export function choice(rnd: () => number, object: { [key: string]: number }) {
   let total = 0;
-  Object.entries(obj).forEach(([_, p]) => (total += p));
+  Object.entries(object).forEach(([_, p]) => (total += p));
   if (total <= 0) throw new Error("Error 62");
   const val = rnd() * total;
   let past = 0;
-  for (let k in obj) {
-    past += obj[k];
+  for (let k in object) {
+    past += object[k];
     if (past >= val) return k;
   }
   throw new Error("Error 70");
